meson: include pre-generated man pages & HTML docs in dist tarballs
iputils is a core utility (hard to avoid on e.g. servers or small images, even)
so it's handy to have a minimal set of build dependencies. Pre-generate
the man pages and HTML docs and include them in the dist tarball so they can
be used without installing libxslt, docbook, and friends.

This is a followup to 83f011ab2a1e8ebf6f6e9b00494ceb069ac42a63 and
72a36873cc909abe51d3d57053fb786402ef734a.

Bug: https://github.com/iputils/iputils/issues/479
Bug: https://bugs.gentoo.org/908817
Bug: https://bugs.gentoo.org/920901
Signed-off-by: Sam James <sam@gentoo.org>