meson: include pre-generated man pages & HTML docs in dist tarballs
iputils is a core utility (hard to avoid on e.g. servers or small images, even)
so it's handy to have a minimal set of build dependencies. Pre-generate
the man pages and HTML docs and include them in the dist tarball so they can
be used without installing libxslt, docbook, and friends.

Create new script build-aux/meson-build-dist-man.sh, which is invoked by
meson itself on 'meson dist' (which is invoked by tools/create-tarballs.sh).

For the reason of the implementation see
https://github.com/mesonbuild/meson/issues/2166 and more specifically,
https://github.com/mesonbuild/meson/issues/2166#issuecomment-629696911.

Closes: https://github.com/iputils/iputils/pull/520
Fixes: https://github.com/iputils/iputils/issues/479
Fixes: https://bugs.gentoo.org/908817
Fixes: https://bugs.gentoo.org/920901
Reviewed-by: Petr Vorel <pvorel@suse.cz>
Signed-off-by: Sam James <sam@gentoo.org>
[ pvorel: unset CC to be able to run make dist on cross compile ]
Signed-off-by: Petr Vorel <pvorel@suse.cz>