ci: Run meson dist
This requires to change permissions for the directory to avoid git error:

    fatal: detected dubious ownership in repository at '/__w/iputils/iputils'
    To add an exception for this directory, call:

	    git config --global --add safe.directory /__w/iputils/iputils
    Error: Process completed with exit code 128.

Only an old git 1.8.3.1 from CentOS 7 does not require that.

Skip running on Ubuntu Bionic which meson 0.45.1 is too old, thus it
does not support '-C builddir' parameter of 'meson dist'. Ubuntu Bionic
is EOL since 04/2023, therefore not fixing it.

Signed-off-by: Petr Vorel <pvorel@suse.cz>