ping: Fix EMSGSIZE on -s > 65527 on ICMP datagram socket
sendto() fails on on -s > 65527 due kernel limitation (both IPv4 and
IPv6 on ICMP datagram socket, IPv4 on raw socket).

But when this size is combined with -c, it prolongs shutdown after send
- it keeps running recvmsg() until timeout. This was fixed by this
change. Maybe we should quit on first EMSGSIZE, because it does not make
much sense to get the same error more times.

This will be workarounded with later lower down -s max size in next
commits, but it's better to fix it before in case the max -s size is
reenabled in the future.

Reproducible with:

    # sysctl -w net.ipv4.ping_group_range="0 2147483647" # ICMP datagram socket
    $ ./builddir/ping/ping -v -i0.2 -c2 -s 65528 127.0.0.1
    $ ./builddir/ping/ping -v -i0.2 -c2 -s 65528 ::1

    # sysctl -w net.ipv4.ping_group_range="1 0" # raw socket socket
    $ ./builddir/ping/ping -v -i0.2 -c2 -s 65528 127.0.0.1

Reviewed-by: Cyril Hrubis <chrubis@suse.cz>
Reviewed-by: Benjamin Poirier <benjamin.poirier@gmail.com>
Signed-off-by: Petr Vorel <pvorel@suse.cz>