meson: fix installation of prebuilt manpages in dist tarball
'page' contains the name of the program we are building/installing a
manual for. In the custom targets for building the manpages, we append
.xml and .8 in turn -- we don't replace a ".xml" in the page name.

As there is nothing to replace, we end up checking if e.g. "./arping"
exists, and trying to install it as a manual if so -- rather than
arping.8 as intended. Sync the logic so that this works.

Fixes: 535f6de ("meson: include pre-generated man pages & HTML docs in dist tarballs")
Closes: https://github.com/iputils/iputils/pull/556
Reviewed-by: Petr Vorel <pvorel@suse.cz>
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>