build.sh: Fix tarball name
'meson dist' produces tarball name based on the version in meson.build,
not on the date of run.

Using meson tools is safer than parsing meson.build with sed, but it
requires to add install jq.

For getting the latest version cannot be used much simpler 'git describe
--abbrev=0' (the latest tag), because fork might have fewer tags or even
no tag.

Fixes: 2e8b3b2 ("build.sh: Add support for meson dist")
Fixes: 6d1f6e0 ("build.sh: Fix testing release build")
Signed-off-by: Petr Vorel <pvorel@suse.cz>