Revert "ping: support IPv4-Mapped-in-IPv6 target addresses"
This reverts commit 8ed7ffc999e2a541e06ee48faf26a323dfe487c2.

This is an incomplete implementation. While there are two attempts
(PR#562 [1] PR#562 [2] and PR#563 [3]) to fix it, it's probably not a
good idea to support IPv4-Mapped-in-IPv6 target addresses.

There were concerns [4] about supporting this when discussion on BusyBox
mailing list about patch bringing this functionality to BusyBox ping
implementation:

* ping is a low-level diagnostic tool, maybe it's not good to assume
  things like this, because IPv4-mapped IPv6 address format is not
  supported by raw sockets
* it's not clear how -4 -6 switches should behave

NOTE: next commit fixes IPv4-Mapped-in-IPv6 target addresses via -4
flag.

Because also fping did not want to support it [6], let's revert the
feature.

[1] https://github.com/iputils/iputils/pull/561
[2] https://github.com/iputils/iputils/pull/563
[3] https://github.com/iputils/iputils/pull/562
[4] https://lists.busybox.net/pipermail/busybox/2024-October/090975.html
[5] https://lists.busybox.net/pipermail/busybox/2024-October/090985.html
[6] https://github.com/schweikert/fping/pull/356

Closes: https://github.com/iputils/iputils/pull/567
Reviewed-by: Ronan Pigott <ronan@rjp.ie>
Signed-off-by: Petr Vorel <pvorel@suse.cz>