Refactor JSON errors
Avoid complicated "splitting" of errors into types, simply return a
"error" key with the message which would have been printed to stderr in
non-JSON output.
Add helper macros to avoid duplication of logic, and to handle "errors"
which are only printed and not raised.
Some errors were purposefully not switched to the new functionality as
they are in difficult to test code paths.
Also split output logic into a separate header file, as it is not
specific to JSON.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>