ping: Fix ipv4 ttl value when using SOCK_DGRAM on big endian systems
7e7ffff attempted to fix a GCC warning about strict aliasing (which
it seems may have been an erroneous one in the first place), but
caused the ttl value when pinging an ipv4 address using SOCK_DGRAM
on a big endian system (for ex. IBM S390) to always appear as 0.
Using memcpy instead of directly casting the value should be the
safest option, fixing the issue and also avoiding the possibility
of unaligned access to the value returned by CMSG_DATA.

Signed-off-by: Radoslav Kolev <radoslav.kolev@suse.com>