ping: Fix regression in -c1
We change the rts->lingertime to be stored directly in us instead of ms
in order to simplify the overflow checks, however we missed two places
where the value was not properly converted.

The initialization of the lingertime has to be converted to us as well
since the MAXWAIT is in seconds it has to be multiplied by 1000000 now.

The check againts the nores_interval has to be updated too since the
nores_interval is initialized from rts->interval that is stored in ms
and the oom_count is a counter, hence the product of the multiplication
is still in ms and the rts->lingertime has to be divided by 1000 in this
case to be converted back to ms.

Fixes: f7d1989 ("ping: Fix integer overflow in large -W value")
Fixes: https://github.com/iputils/iputils/issues/596
Closes: https://github.com/iputils/iputils/pull/597
Reported-by: Alberto Salvia Novella <es20490446e.wordpress.com>
Co-developed-by: Petr Vorel <pvorel@suse.cz>
Signed-off-by: Cyril Hrubis <chrubis@suse.cz>