tracepath: Add raw extension data field to struct probehdr
The 'probehdr' structure describes the contents of the UDP probe payload
used by tracepath.

ICMP extensions, which will be introduced in the upcoming patches, are read
from the socket's error queue together with the UDP payload, and are
appended to the end of the probe.

To support these extensions and to allow easier parsing of raw ICMP
payloads received from the socket's error queue, add a new raw-data byte
array to 'probehdr' and increase its buffer size.
Since the maximum size of ICMPv4 error messages is 576 bytes and the
maximum size of ICMPv6 error messages is 1280 bytes, allocate 1280 bytes to
this buffer so both cases are covered.

Signed-off-by: Danielle Ratson <danieller@nvidia.com>